requirements.txt
pymongo
netifaces
psutil
python-nmap
psycopg2-binary
